+++ src/distributions/float.rs
@@ -148,7 +148,7 @@ float_impls! { f64, u64, 52, 1023 }
 mod tests {
     use Rng;
     use distributions::{Open01, OpenClosed01};
-    use mock::StepRng;
+    use rngs::mock::StepRng;
 
     const EPSILON32: f32 = ::core::f32::EPSILON;
     const EPSILON64: f64 = ::core::f64::EPSILON;
